+++ Dockerfile.test_tcp_example
@@ -10,7 +10,7 @@ FROM micropython/unix:v1.18
 # use "volumes" in docker-compose file to remove need of rebuilding
 # COPY ./ /home
 # COPY umodbus /root/.micropython/lib/umodbus
-# COPY unittest.py /root/.micropython/lib/unittest.py
+# COPY mpy_unittest.py /root/.micropython/lib/mpy_unittest.py
 
 RUN micropython-dev -m upip install micropython-ulogging
 RUN micropython-dev -m upip install micropython-urequests
