+++ Dockerfile.tests
@@ -11,10 +11,10 @@ COPY ./ /home
 # keep examples and tests registers JSON file easily in sync
 COPY registers/example.json /home/tests/test-registers.json
 COPY umodbus /root/.micropython/lib/umodbus
-COPY unittest.py /root/.micropython/lib/unittest.py
+COPY mpy_unittest.py /root/.micropython/lib/mpy_unittest.py
 
 RUN micropython-dev -m upip install micropython-ulogging
 RUN micropython-dev -m upip install micropython-urequests
-RUN micropython-dev -c "import unittest; unittest.main('tests')"
+RUN micropython-dev -c "import mpy_unittest as unittest; unittest.main('tests')"
 
 ENTRYPOINT ["/bin/bash"]
