+++ tests/test_const.py
@@ -4,7 +4,7 @@
 
 from umodbus.typing import List
 import ulogging as logging
-import unittest
+import mpy_unittest as unittest
 from umodbus import const as Const
